Move server/hub.go -> server/websocket.go
